AI_Persona/backend/requirements.txt
livekit-agents[openai,elevenlabs]>=1.3.0
python-dotenv>=1.0.0
